add test and initial handling switch for invoke for GVM calls.
